aws_cv_clip/src/requirements.txt
# AWS 다이어그램 오토라벨러 의존성 패키지

# 딥러닝 프레임워크
torch>=1.12.0
torchvision>=0.13.0

# CLIP 모델
open-clip-torch>=2.0.0

# 컴퓨터 비전
opencv-python>=4.6.0
Pillow>=9.0.0

# 수치 계산
numpy>=1.21.0

# 벡터 검색
faiss-cpu>=1.7.0
# GPU 사용 시: faiss-gpu>=1.7.0

# 진행률 표시
tqdm>=4.64.0

# 데이터 처리
pandas>=1.4.0

# 설정 파일
PyYAML>=6.0

# OCR
easyocr>=1.6.0

# 문자열 매칭
rapidfuzz>=2.0.0

# 이미지 처리 (선택사항)
scikit-image>=0.19.0

# 로깅 (선택사항)
loguru>=0.6.0

# 타입 힌트 (개발용)
typing-extensions>=4.0.0

# 테스트 (개발용)
pytest>=7.0.0
pytest-cov>=3.0.0
